+++ module/actor/actor.mjs
@@ -1886,7 +1886,13 @@ export class HeroSystem6eActor extends Actor {
             }
 
             // Rollable Characteristics
-            this.updateRollable(key.toLowerCase());
+            const rollableChanges = this.updateRollable(key.toLowerCase());
+            if (rollableChanges) {
+                changed = true;
+
+                // TODO: FIXME: Not quite right.
+                foundry.utils.mergeObject(changes, rollableChanges);
+            }
         }
 
         // Save changes
@@ -2034,12 +2040,18 @@ export class HeroSystem6eActor extends Actor {
             xmlid: key.toUpperCase(),
             actor: this,
         });
+
         if (characteristic && charPowerEntry?.behaviors.includes("success")) {
             characteristic.roll = Math.round(9 + characteristic.value * 0.2);
             if (!this.system.is5e && characteristic.value < 0) {
                 characteristic.roll = 9;
             }
+            return {
+                [`system.characteristics.${key}.roll`]: characteristic.roll,
+            };
         }
+
+        return undefined;
     }
 
     async CalcActorRealAndActivePoints() {